pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "folder-tokenizer"
version = "0.1.0"
description = "Analyze token counts for all documents in a folder using HuggingFace tokenizers"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Sajal Garg", email = "sajalgargciv2805@gmail.com"}
]
dependencies = [
    "transformers>=4.30.0",
    "python-docx>=0.8.11",
    "pypdf>=3.0.0",
    "pytesseract>=0.3.10",
    "Pillow>=9.0.0",
    "chardet>=5.0.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
ui = [
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
folder-tokenizer = "folder_tokenizer.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
